guessmovie.py
This is a game created using simple Python code. The program's objective is for the user to guess the name of a movie. Initially, the program generates a series of stars representing the number of characters in the movie title. The player is then prompted to guess letters that may be in the movie title. If the guessed letter is correct, the program updates the display by replacing the stars with the correct letter. 

At any point in the game, if the player feels confident, they can choose to answer the entire movie title. They will be prompted to press 1 to submit their answer or 0 to unlock one more character. The game continues with both players participating, and when a player wants to end the game, they can press 0 to exit. The program will then display the points accumulated by each player.