Create fizz_buzz.py
This is a simple Python program that takes an input (n) from the user and generates a sequence of numbers up to that range (n). The program checks each number to determine if it is divisible by 3 or 5. If a number is divisible by 3, it prints "Fizz" along with the corresponding number. If a number is divisible by 5, it prints "Buzz" along with the number. If a number is divisible by both 3 and 5, it prints "FizzBuzz" along with the number. If the number is not divisible by either, it simply prints the number itself.